ConfluentSchemaBasedMessagesSerdeProvider cleanup
