+++ docs/Changelog.md
@@ -33,7 +33,7 @@
 * [#3238](https://github.com/TouK/nussknacker/pull/3238) K8 runtime's logback conf can be stored in single ConfigMap for all runtime pods
 * [#3201](https://github.com/TouK/nussknacker/pull/3201) Added literal types
 * [#3240](https://github.com/TouK/nussknacker/pull/3240) Error topic created by default if not exists
-* [#3245](https://github.com/TouK/nussknacker/pull/3245) [#3265](https://github.com/TouK/nussknacker/pull/3265) [#3288](https://github.com/TouK/nussknacker/pull/3288) Universal kafka source/sink, handling multiple scenarios like: avro message for avro schema, json message for json schema 
+* [#3245](https://github.com/TouK/nussknacker/pull/3245) [#3265](https://github.com/TouK/nussknacker/pull/3265)[#3288](https://github.com/TouK/nussknacker/pull/3288)[#3295](https://github.com/TouK/nussknacker/pull/3295) Universal kafka source/sink, handling multiple scenarios like: avro message for avro schema, json message for json schema 
 * [#3249](https://github.com/TouK/nussknacker/pull/3249) Confluent 5.5->7.2, avro 1.9->1.11 bump
 * [#3250](https://github.com/TouK/nussknacker/pull/3250) Kafka 2.4 -> 3.2, flink 0.14.4 -> 0.14.5
 * [#3270](https://github.com/TouK/nussknacker/pull/3270) Added type representing null
